Focus graph surgery unit tests on individual surgeons
Remove redundant multi-surgery pipeline cases and retain single-surgeon contracts plus the registry smoke test. Split RoPE domain checks into independent parameterized cases and cover unrelated-node rejection directly, preserving the exact set of covered production lines.
